refactor(deno/postgres-js): use DSQL postgres.js connector
Replace the raw deno-postgres driver and hand-written IAM token
generator with the official @aws/aurora-dsql-postgresjs-connector.
Rename deno/deno-postgres/ to deno/postgres-js/ to match the other
postgres.js sample in the repo.
