+++ README.md
@@ -31,7 +31,7 @@ The subdirectories contain code examples for connecting and using Aurora DSQL in
 | Typescript  |               [Prisma](typescript/prisma)               |
 | Typescript  |            [Sequelize](typescript/sequelize)            |
 | Typescript  |             [TypeORM](typescript/type-orm)              |
-|    Deno     |        [deno-postgres](deno/deno-postgres/)              |
+|    Deno     |        [postgres-js](deno/postgres-js/)              |
 
 
 |  Language   |                 Cluster Management                  |
@@ -61,7 +61,7 @@ The subdirectories contain code examples for connecting and using Aurora DSQL in
 |                          Sample Applications                          |
 |:---------------------------------------------------------------------:|
 | [Amazon Aurora DSQL Agent](sample-amazon-aurora-dsql-agent) — A sample application showing how to build an AI agent that queries Aurora DSQL using natural language, powered by Strands Agents, Bedrock AgentCore Gateway (MCP), and AgentCore Runtime (A2A) |
-| [Booking API (Deno)](deno/booking-api/deno-postgres) — A booking/reservation REST API built with Deno.serve() and deno-postgres, demonstrating OCC retry handling, IAM authentication, and serverless-ready architecture |
+| [Booking API (Deno)](deno/booking-api/postgres-js) — A booking/reservation REST API built with Deno.serve() and postgres.js (via the Aurora DSQL connector), demonstrating OCC retry handling, unique-window enforcement, IAM authentication, and serverless-ready architecture |
 
 Each example includes language and client-specific instructions as well as instructions to invoke example code.
