--- deno/deno-postgres/src/example_preferred.ts
@@ -1,185 +1,0 @@
-/**
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- *
- * Aurora DSQL with deno-postgres — Preferred Example
- *
- * Demonstrates how to connect to an Aurora DSQL cluster from Deno using the
- * native deno-postgres driver (jsr:@db/postgres). This example uses IAM token
- * authentication via @aws-sdk/dsql-signer and connects over SSL with
- * verify-full mode.
- *
- * Unlike Node.js samples, Deno runs TypeScript natively — no build step,
- * no node_modules, no package.json. Just `deno task start`.
- *
- * Prerequisites:
- *   - Deno 2.x installed
- *   - AWS credentials configured (default credential chain)
- *   - An active Aurora DSQL cluster
- *
- * Environment variables:
- *   CLUSTER_ENDPOINT - Aurora DSQL cluster endpoint
- *                      (e.g., "foo0bar1baz2quux3quuux4.dsql.us-east-1.on.aws")
- *   CLUSTER_USER     - Database user (e.g., "admin")
- *
- * Usage:
- *   export CLUSTER_ENDPOINT="<your endpoint>"
- *   export CLUSTER_USER="admin"
- *   deno task start
- */
-
-import { Client } from "@db/postgres";
-import { DsqlSigner } from "@aws-sdk/dsql-signer";
-
-const NUM_CONCURRENT_QUERIES = 8;
-
-/**
- * Generates a short-lived IAM authentication token for Aurora DSQL.
- *
- * Aurora DSQL does not use traditional passwords. Instead, it uses IAM tokens
- * generated by @aws-sdk/dsql-signer. Tokens are scoped to a specific cluster
- * endpoint and expire after a short period.
- *
- * The region is resolved from (in priority order):
- *   1. The `AWS_REGION` environment variable (standard AWS convention)
- *   2. `"us-east-1"` (safe fallback for samples)
- *
- * Extracting the region from the endpoint hostname is avoided — the hostname
- * format is an implementation detail that can change silently without an
- * SDK bump.
- *
- * @param clusterEndpoint - The Aurora DSQL cluster hostname
- * @param isAdmin - Whether to generate an admin token (default: true)
- * @returns A short-lived authentication token string
- */
-async function generateToken(
-  clusterEndpoint: string,
-  isAdmin = true,
-): Promise<string> {
-  const region = Deno.env.get("AWS_REGION") ?? "us-east-1";
-
-  const signer = new DsqlSigner({
-    hostname: clusterEndpoint,
-    region,
-  });
-
-  return isAdmin
-    ? await signer.getDbConnectAdminAuthToken()
-    : await signer.getDbConnectAuthToken();
-}
-
-/**
- * Creates a deno-postgres Client connected to Aurora DSQL.
- *
- * Key DSQL-specific configuration:
- *   - Port 5432 (PostgreSQL default)
- *   - Database is always "postgres" (DSQL creates this automatically)
- *   - TLS is mandatory with verify-full (Deno's built-in TLS verifies
- *     against the system CA store by default)
- *   - Password is a short-lived IAM token, not a static credential
- *
- * @param clusterEndpoint - The Aurora DSQL cluster hostname
- * @param user - The database user name
- * @returns A connected deno-postgres Client
- */
-async function createClient(
-  clusterEndpoint: string,
-  user: string,
-): Promise<Client> {
-  const isAdmin = user === "admin";
-  const token = await generateToken(clusterEndpoint, isAdmin);
-
-  const client = new Client({
-    hostname: clusterEndpoint,
-    port: 5432,
-    user: user,
-    password: token,
-    database: "postgres",
-    // Aurora DSQL requires SSL with verify-full.
-    // deno-postgres uses Deno's built-in TLS which verifies against
-    // the system CA certificate store by default.
-    tls: {
-      enabled: true,
-      enforce: true,
-    },
-  });
-
-  await client.connect();
-  return client;
-}
-
-/**
- * Runs a single worker query to verify the connection pool works.
- * Each worker executes a simple SELECT and asserts the result.
- */
-async function worker(client: Client, workerId: number): Promise<void> {
-  const result = await client.queryObject<{ worker_id: number }>(
-    "SELECT $1::int as worker_id",
-    [workerId],
-  );
-  console.log(`Worker ${workerId} result: ${result.rows[0].worker_id}`);
-
-  if (result.rows[0].worker_id !== workerId) {
-    throw new Error(
-      `Worker ${workerId}: expected ${workerId}, got ${result.rows[0].worker_id}`,
-    );
-  }
-}
-
-/**
- * Main example function.
- *
- * Connects to Aurora DSQL, runs concurrent queries to verify connectivity,
- * and cleans up. This follows the same pattern as the JavaScript
- * node-postgres example in this repository.
- */
-async function example(): Promise<void> {
-  const clusterEndpoint = Deno.env.get("CLUSTER_ENDPOINT");
-  if (!clusterEndpoint) {
-    console.error("CLUSTER_ENDPOINT environment variable is not set");
-    Deno.exit(1);
-  }
-
-  const user = Deno.env.get("CLUSTER_USER");
-  if (!user) {
-    console.error("CLUSTER_USER environment variable is not set");
-    Deno.exit(1);
-  }
-
-  // Note: deno-postgres does not have a built-in connection pool like node-postgres.
-  // For concurrent queries, we create multiple client connections.
-  // In production, consider using a pool wrapper or managing connections manually.
-  const clients: Client[] = [];
-
-  try {
-    // Create connections and run concurrent queries
-    const workers: Promise<void>[] = [];
-    for (let i = 1; i <= NUM_CONCURRENT_QUERIES; i++) {
-      const client = await createClient(clusterEndpoint, user);
-      clients.push(client);
-      workers.push(worker(client, i));
-    }
-
-    // Wait for all workers to complete
-    await Promise.all(workers);
-
-    console.log(
-      "Connection pool with concurrent connections exercised successfully",
-    );
-  } catch (error) {
-    console.error(error);
-    throw error;
-  } finally {
-    // Close all connections
-    for (const client of clients) {
-      await client.end();
-    }
-  }
-}
-
-export { example };
-
-// Run the example when executed directly (not when imported by tests)
-if (import.meta.main) {
-  example();
-}
