+++ deno/postgres-js/test/example_preferred.test.ts
@@ -2,7 +2,7 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
  *
- * Integration test for the deno-postgres preferred example.
+ * Integration test for the postgres.js preferred example.
  *
  * Requires CLUSTER_ENDPOINT and CLUSTER_USER environment variables.
  * Skips if not set (no DSQL cluster available).
